% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary-functions-visuals.R
\name{annotationBarplot}
\alias{annotationBarplot}
\title{add annotation to AllelicImbalance barplot}
\usage{
annotationBarplot(strand, snp, lowerLeftCorner, annDfPlus, annDfMinus,
  cex = 0.7, ypos = 0, interspace = 1)
}
\arguments{
\item{strand}{strand, "+", "-", "*" or "both"}

\item{snp}{integer for the described snp}

\item{lowerLeftCorner}{position of the plot to add legend to (default c(0,0))}

\item{annDfPlus}{annotation dataframe plus strand}

\item{annDfMinus}{annotation dataframe minus strand}

\item{cex}{size of annotation text}

\item{ypos}{relative y-axis position for the annotation text}

\item{interspace}{space between each annotation block}
}
\description{
adds a customizable annotation functionality for AllelicImbalance barplots.
}
\details{
the function is preferably called from within the AllelicImbalance barplot method.
}
\examples{

#code placeholders
#< create a barplot without annotation >
#< add annotation >
 
}
\author{
Jesper R. Gadin
}
\keyword{annotation}
\keyword{barplot}

