% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{ASEset-class}
\alias{ASEset}
\alias{ASEset-class}
\alias{alleleCounts}
\alias{alleleCounts,ASEset-method}
\alias{arank}
\alias{arank,ASEset-method}
\alias{fraction}
\alias{fraction,ASEset-method}
\alias{mapBias}
\alias{mapBias,ASEset-method}
\title{ASEset objects}
\usage{
alleleCounts(x, strand = "nonStranded")

mapBias(x)

fraction(x, strand = "nonStranded", verbose = FALSE)

arank(x, ret = "names", strand = "nonStranded", ...)
}
\arguments{
\item{x}{ASEset object}

\item{strand}{which strand of 'nonStranded', '+', '-' or '*'}

\item{verbose}{makes function more talkative}

\item{ret}{return names or counts}

\item{...}{additional arguments}
}
\value{
An object of class ASEset containing location information and allele
counts for a number of SNPs measured in a number of samples on various
strand, as well as mapBias information. All data is stored in a manner
similar to the \code{\link[GenomicRanges]{SummarizedExperiment}} class.
}
\description{
Object that holds allele counts, genomic positions and map-bias for a set of
SNPs
}
\details{
An ASEset object differs from a regular SummarizedExperiment object in that
the assays contains an array instead of matrix. This array has ranges on the
rows, sampleNames on the columns and variants in the third dimension.

It is possible to use the commands barplot and locationplot on an ASEset
object see more details in \code{\link{barplot}} and
\code{\link{locationplot}}.

Four different alleleCount options are available. The simples one is the
nonStranded option, and is experiments where the strand information is not
known. In this option both the plus, minus and unknown strand will be
counted and present. The unknown strand is when the aligner could not find
any strand associated with the read. Then there are an option too add plus
and minus stranded data. When using this, it is essential to make sure that
the RNA-seq experiment under analysis has in fact been created so that
correct strand information was obtained.
}
\section{Constructor}{
 ASEsetFromCountList(rowData, countListNonStranded =
NULL, countListPlus = NULL, countListMinus = NULL, countListUnknown = NULL,
colData = NULL, mapBiasExpMean = array(),verbose=FALSE ...)

\describe{

Arguments: \item{rowData}{A \code{GenomicRanges object} that contains the
variants of interest} \item{countListNonStranded}{A \code{list} where each
entry is a matrix with allele counts as columns and sample counts as rows}
\item{countListPlus}{A \code{list} where each entry is a matrix with allele
counts as columns and sample counts as rows} \item{countListMinus}{A
\code{list} where each entry is a matrix with allele counts as columns and
sample counts as rows} \item{countListUnknown}{A \code{list} where each
entry is a matrix with allele counts as columns and sample counts as rows}
\item{colData}{A \code{DataFrame} object containing sample specific data}
\item{mapBiasExpMean}{A 3D \code{array} describing mapping bias. The SNPs
are in the 1st dimension, samples in the 2nd dimension and variants in the
3rd dimension.} \item{verbose}{Makes function more talkative}
\item{...}{arguments passed on to SummarizedExperiment constructor} }
}
\examples{
\\t#make example countList
\\tset.seed(42)
\\tcountListPlus <- list()
\\tsnps <- c('snp1','snp2','snp3','snp4','snp5')
\\tfor(snp in snps){
\\t  count<-matrix(rep(0,16),ncol=4,dimnames=list(
\\t\\tc('sample1','sample2','sample3','sample4'),
\\t\\tc('A','T','G','C')))
\\t  #insert random counts in two of the alleles
\\t  for(allele in sample(c('A','T','G','C'),2)){
\\t\\tcount[,allele]<-as.integer(rnorm(4,mean=50,sd=10))
\\t  }
\\t  countListPlus[[snp]] <- count
\\t}

\\t#make example rowData
\\trowData <- GRanges(
\\t  seqnames = Rle(c('chr1', 'chr2', 'chr1', 'chr3', 'chr1')),
\\t  ranges = IRanges(1:5, width = 1, names = head(letters,5)),
\\t  snp = paste('snp',1:5,sep='')
\\t)
\\t#make example colData
\\tcolData <- DataFrame(Treatment=c('ChIP', 'Input','Input','ChIP'),
\\t \\trow.names=c('ind1','ind2','ind3','ind4'))

\\t#make ASEset
\\ta <- ASEsetFromCountList(rowData, countListPlus=countListPlus,
\\tcolData=colData)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The
\code{\link[GenomicRanges]{SummarizedExperiment}} for ranges operations.  }
}
\keyword{ASEset}
\keyword{class}

