% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/barplot-methods.R
\docType{methods}
\name{ASEset-topalleleexample}
\alias{ASEset-topalleleexample}
\alias{topalleleexample}
\alias{topalleleexample,ASEset-method}
\title{topalleleexample ASEset objects}
\usage{
topalleleexample(x)
}
\arguments{
\item{x}{"missing", and so no argument is given}
}
\description{
used as a visual example to the top allele criteria
}
\details{
shows what is top and what is down
}
\examples{
data(ASEset)
topalleleexample(ASEset[1])
}
\author{
Jesper R. Gadin
}
\keyword{topallele}

