% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllelicImbalance-package.R
\docType{data}
\name{ASEset.old}
\alias{ASEset.old}
\title{ASEset.old object}
\description{
old version of an ASEset which needs to be updated
}
\examples{
##load eample data (Not Run)
#data(ASEset.old)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{data}
\keyword{example}
\keyword{object}

