% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllelicImbalance-package.R
\docType{data}
\name{ASEset.sim}
\alias{ASEset.sim}
\title{ASEset.sim object}
\description{
ASEset with simulated data with SNPs within the first 200bp of chromosome 17,
which is required to have example data for the refAllele function.
}
\examples{
##load eample data (Not Run)
#data(ASEset.sim)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{data}
\keyword{example}
\keyword{object}

