% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxillary-methods.R
\docType{methods}
\name{ASEsetFromOldVersion}
\alias{ASEsetFromOldVersion}
\alias{ASEsetFromOldVersion,ASEset-method}
\title{update ASEset}
\usage{
ASEsetFromOldVersion(x, ...)

\S4method{ASEsetFromOldVersion}{ASEset}(x)
}
\arguments{
\item{x}{ASEset object}

\item{...}{passed on to ASEsetFromOldVersion function}
}
\description{
converts old ASEsets to new ASEset objects
}
\details{
It is not possible to convert back to an old version.
}
\examples{
#load an old example object
data(ASEsetOld1)

# convert to new version
a <- ASEsetFromOldVersion(ASEsetOld1)
}
\author{
Jesper R. Gadin
}
\keyword{ASEset}

