% Generated by roxygen2 (4.0.1): do not edit by hand
\name{barplot-lattice-support}
\alias{barplot-lattice-support}
\alias{barplotLatticeCounts}
\alias{barplotLatticeFraction}
\title{lattice barplot inner functions for ASEset objects}
\usage{
barplotLatticeFraction(identifier, afraction, arank, amainVec, ...)

barplotLatticeCounts(identifier, acounts, arank, amainVec, ...)
}
\arguments{
\item{identifier,}{the single snp name to plot}

\item{afraction}{the fractions in matrix for each snp}

\item{arank}{the ranks for each snp}

\item{acounts}{the counts for each snp}

\item{amainVec,}{vector for each plots main}

\item{...}{for simpler generics when extending function}
}
\description{
Generates lattice barplots for ASEset objects. Two levels of plotting detail
are provided: a detailed barplot of read counts by allele useful for fewer
samples and SNPs, and a less detailed barplot of the fraction of imbalance,
useful for more samples and SNPs.
}
\details{
\code{filter.pValue.fraction} is intended to remove p-value annotation with
very large difference in frequency, which could just be a sequencing
mistake. This is to avoid p-values like 1e-235 or similar.

\code{sampleColour}User specified colours, either given as named colours
('red', 'blue', etc) or as hexadecimal code. Can be either length 1 for all
samples, or else of a length corresponding to the number of samples for
individual colouring.
}
\examples{
\\ta <- ASEset
\\tstrand <- '+'
\\tacounts <-  alleleCounts(a,strand=strand)
\\tarank <-  arank(a,strand=strand)
\\tafraction <- fraction(a, strand=strand)
\\tamainVec <- rep('',nrow(a))

\\tname <- rownames(a)[1]

\\tbarplotLatticeFraction(identifier=name, afraction, arank, amainVec)
\\tbarplotLatticeCounts(identifier=name,  acounts, arank, amainVec)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The \code{\link{ASEset}} class which the barplot
function can be called up on.  }
}
\keyword{barplot}

