% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicimbalance.R
\name{beta_binom_pval}
\alias{beta_binom_pval}
\title{p-value for a beta-binomial variable}
\usage{
beta_binom_pval(q, size, shape1, shape2)
}
\arguments{
\item{q}{vector of quantiles}

\item{size}{vector of sample size parameters}

\item{shape1, shape2}{the two (positive) shape parameters of the standard
beta distribution. See the documentation for \code{Betabinom} in the
\code{VGAM} package.}
}
\value{
numeric. The p-value.
}
\description{
Hypothesis test against a beta-binomial null.
}
