% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{binom.test}
\alias{binom.test}
\alias{binom.test,ASEset-method}
\title{binomial test}
\usage{
binom.test(x, n, p = 0.5, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95)

\S4method{binom.test}{ASEset}(x, n = "*")
}
\arguments{
\item{x}{\linkS4class{ASEset} object}

\item{n}{strand option}
}
\value{
\code{binom.test} returns a matrix
}
\description{
Performs a binomial test on an ASEset object.
}
\details{
the test can only be applied to one strand at the time.
}
\examples{
#load example data
data(ASEset)

#make a binomial test
binom.test(ASEset,'*')
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The chisq.test which is another test
that can be applied on an \linkS4class{ASEset} object.  }
}
\keyword{binomial}
\keyword{test}

