% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxillary-methods.R
\docType{methods}
\name{defaultPhase}
\alias{defaultPhase}
\alias{defaultPhase,numeric-method}
\title{defaultPhase}
\usage{
defaultPhase(i, ...)

\S4method{defaultPhase}{numeric}(i, j, ...)
}
\arguments{
\item{i}{number of rows}

\item{...}{arguments to forward to internal functions}

\item{j}{number of columns}
}
\description{
used to populate the phase slot in an ASEset object
}
\details{
will set everything to 0
}
\examples{
i <- 5
j <- 10
defaultPhase(i,j)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{phase}

