% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{detectAI}
\alias{detectAI}
\alias{detectAI,ASEset-method}
\title{detectAI}
\usage{
detectAI(x, ...)

\S4method{detectAI}{ASEset}(x, return.class = "logical",
  return.type = "all", strand = "*", threshold.frequency = 0.05,
  threshold.count.sample = 5, min.delta.frequency = 0.05,
  max.pvalue = 0.05, function.test = "binom.test")
}
\arguments{
\item{x}{ASEset}

\item{...}{internal arguments}

\item{return.class}{class to return (atm only class 'logical')}

\item{return.type}{'ref' ,'alt' or default: 'all'}

\item{strand}{strand to infer from}

\item{threshold.frequency}{least fraction to classify (see details)}

\item{threshold.count.sample}{least amount of counts to try to infer allele}

\item{min.delta.frequency}{minimum of frequency difference from 0.5 (or mapbias adjusted value)}

\item{max.pvalue}{pvalue over this number will be filtered out}

\item{function.test}{At the moment the only available option is 'binomial.test'}
}
\description{
detection of AllelicImbalance
}
\details{
threshold.frequency is the least fraction needed to classify as bi tri or
quad allelic SNPs. If 'all' then all of bi tri and quad allelic SNPs will use the same
threshold. Everything under the treshold will be regarded as noise. 'all' will return
a matrix with snps as rows and uni bi tri and quad will be columns. For this function
Anything that will return TRUE for tri-allelicwill also return TRUE for uni and bi-allelic
for the same SNP an Sample.

return.type 'ref' return only AI when reference allele is more expressed. 'alt' return only
AI when alternative allele is more expressed or 'all' for both 'ref' and 'alt' alleles.
Reference allele is the one present in the reference genome on the forward strand.

min.delta.frequency and function.test will use the value in mapBias(x) as expected value.

function.test will use the two most expressed alleles for testing. Make therefore sure there
are no tri-allelic SNPs or somatic mutations among the SNPs in the ASEset.
}
\examples{
#load example data
data(ASEset)
a <- ASEset

dai <- detectAI(a)
}
\author{
Jesper R. Gadin
}
\keyword{infer}

