% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{DetectedAI-class}
\alias{DetectedAI}
\alias{DetectedAI-class}
\alias{DetectedAI-method}
\title{DetectedAI class}
\arguments{
\item{x}{ASEset object or list of ASEsets}

\item{TxDb}{A \code{transcriptDb} object}

\item{...}{pass arguments to internal functions}
}
\value{
An object of class DetectedAI containing all data to make report.
}
\description{
Object that holds results from a global AI analysis including reference bias
estimations and AI detection.
}
\details{
The DetectedAI-class contains summaries needed to create an AI-report
}
\examples{
data(ASEset)
a <- ASEset
gba <- gba(a)

#summary(gba)
#write.tables(gba)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{ASEset}
\keyword{class}

