% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list-methods.R
\name{general-list}
\alias{general-list}
\alias{list.depth}
\alias{multiUlist.index.names}
\alias{multiUnlist}
\alias{multiUnlist.index}
\alias{multiUnlist.index.names}
\alias{region.list.populate}
\title{general list methods}
\usage{
list.depth(this, thisdepth = 0)

multiUnlist(lst, ...)

multiUnlist.index(lst, expand.lowest.level = FALSE)

multiUnlist.index.names(lst, expand.lowest.level = FALSE)

region.list.populate(ar, idx.mat, idx.mat.names)
}
\arguments{
\item{this}{a list}

\item{thisdepth,}{start value for recursion}

\item{lst,}{the single snp name to plot}

\item{...}{pass on variables internally}

\item{expand.lowest.level}{logical}

\item{ar}{array to use for repopulation}

\item{idx.mat}{array to use for repopulation}

\item{idx.mat.names}{array to use for repopulation}
}
\description{
general list methods that supports flatteing and indexing of lists
}
\details{
This is a collection of supportive list methods. The documentation will
be improved before next release.
}
\examples{
#some example code here
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{list}

