% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{genofilters}
\alias{genofilters}
\alias{hetFilt}
\alias{hetFilt,ASEset-method}
\title{genotype filter methods}
\arguments{
\item{x}{\code{ASEset} object}
}
\description{
useful genotype filters
}
\details{
These filters are called upon ASEset objects
}
\examples{
#load example data
data(ASEset)
a <- ASEset
genotype(a) <- inferGenotypes(a)
hets <- hetFilt(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{filter}

