% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllelicImbalance-package.R
\docType{data}
\name{genomatrix}
\alias{genomatrix}
\title{genomatrix object}
\description{
genomatrix is an example of a matrix with genotypes
}
\examples{
##load eample data (Not Run)
#data(genomatrix)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{data}
\keyword{genotype}

