% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getSnpIdFromLocation}
\alias{getSnpIdFromLocation}
\title{Get rsIDs from locations of SNP}
\usage{
getSnpIdFromLocation(GR, SNPloc, return.vector = FALSE, verbose = TRUE)
}
\arguments{
\item{GR}{A \code{GRanges} that contains positions of SNPs to look up}

\item{SNPloc}{A \code{SNPlocs object} containing information on SNP
locations (e.g. SNPlocs.Hsapiens.dbSNP.xxxxxxxx)}

\item{return.vector}{Setting \code{return.vector=TRUE} returns vector with
rsIds}

\item{verbose}{Setting \code{verbose=TRUE} makes function more talkative}
}
\value{
\code{getSnpIdFromLocation} returns the same GRanges object it was
given with, but with updated with rs.id information.
}
\description{
Given a GRanges object of SNPs and a SNPlocs annotation, this function
attempts to replace the names of the GRanges object entries with rs-IDs.
}
\details{
This function is used to try to identify the rs-IDs of SNPs in a GRanges
object.
}
\examples{
\\t#load example data
\\tdata(ASEset)
\\t
\\t#get counts at the three positions specified in GRvariants
\\tif(require(SNPlocs.Hsapiens.dbSNP.20120608)){
\\t\\tupdatedGRanges<-getSnpIdFromLocation(rowData(ASEset), SNPlocs.Hsapiens.dbSNP.20120608)
\\t\\trowData(ASEset)<-updatedGRanges
\\t}
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{SNP}
\keyword{rs-id}

