% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalAnalysis-class.R
\docType{class}
\name{GlobalAnalysis-class}
\alias{GlobalAnalysis-class}
\alias{GlobalAnalysis}
\alias{GlobalAnalysis-method}
\alias{GlobalAnalysis-class}
\title{GlobalAnalysis class}
\arguments{
\item{x}{ASEset object or list of ASEsets}

\item{TxDb}{A \code{transcriptDb} object}

\item{...}{pass arguments to internal functions}
}
\value{
An object of class GlobalAnalysis containing all data to make report.
}
\description{
Object that holds results from a global AI analysis including reference bias
estimations and AI detection.
}
\details{
The GlobalAnalysis-class contains summaries and "pre-configured and pre-calculated lattice plots" needed to create an AI-report
}
\examples{

data(ASEset)
#a <- ASEset
#gba <- gba(a)

#report(gba)
#write.tables(gba)
#graphs(gba)
#as.list(gba)

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{ASEset}
\keyword{class}
