% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllelicImbalance-package.R
\docType{data}
\name{GRvariants}
\alias{GRvariants}
\title{GRvariants object}
\description{
this data is a \code{GRanges} object that contains the ranges for three
example SNPs.
}
\examples{

#load example data
data(GRvariants)


}
\seealso{
\itemize{ \item The \code{\link{reads}} which is another example
object }
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{data}
\keyword{example}
\keyword{object}
