% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{histplot}
\alias{hist}
\alias{hist,ASEset-method}
\alias{hist,ReferenceBias-method}
\alias{histplot}
\title{histogram plots}
\usage{
\S4method{hist}{ReferenceBias}(x, strand = "*", ...)

\S4method{hist}{ASEset}(x, strand = "*", type = "mean", log = 1, ...)
}
\arguments{
\item{x}{\code{ReferenceBias} object or \code{ASEset} object}

\item{strand}{'+','-' or '*'}

\item{...}{arguments to forward to interal boxplots function}

\item{type}{'mean' (only one option atm)}

\item{log}{an integer to log each value (integer 10 for log10)}
}
\description{
uses base graphics hist plot
}
\details{
The histogram will show the density over frequencies for each sample
}
\examples{
##load example data

#data(ASEset)
#a <- ASEset
#hist(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{hist}
\keyword{plot}

