% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/deprecations.R
\name{implodeList.old}
\alias{implodeList.old}
\title{implode list of arguments into environment}
\usage{
implodeList.old(x)
}
\arguments{
\item{x}{list of variables}
}
\description{
apply on list of variables to be put in the local environment
}
\details{
help the propagation of e.g. graphical paramters
}
\examples{
lst <- list(hungry='yes', thirsty='no')
implodeList.old(lst)
#the check ls()
 ls()
}
\author{
Jesper R. Gadin
}
\keyword{implode}

