% Generated by roxygen2 (4.0.2): do not edit by hand
\name{implodeList}
\alias{implodeList}
\title{implode list of arguments into environment}
\usage{
implodeList(x)
}
\arguments{
\item{x}{list of variables}
}
\description{
apply on list of variables to be put in the local environment
}
\details{
help the propagation of e.g. graphical paramters
}
\examples{
lst <- list(hungry='yes', thirsty='no')
implodeList(lst)
#the check ls()
 ls()
}
\author{
Jesper R. Gadin
}
\keyword{implode}

