% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inferAlleles}
\alias{inferAlleles}
\title{inference of SNPs of ASEset}
\usage{
inferAlleles(x, strand = "*", return.type = "all")
}
\arguments{
\item{x}{ASEset}

\item{strand}{strand}

\item{return.type}{'biallelic' 'triallelic' 'all'}
}
\description{
inference of SNPs
}
\details{
Oftern necessary information to link AI to SNPs outside coding region
}
\examples{
data(ASEset)
g <- inferBiAllelicSnps(ASEset)
}
\author{
Jesper R. Gadin
}
\keyword{infer}

