% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/inference-methods.R
\docType{methods}
\name{inferAltAllele}
\alias{inferAltAllele}
\alias{inferAltAllele,ASEset-method}
\title{inferAltAllele}
\arguments{
\item{x}{\code{matrix} see examples}

\item{return.class}{class of returned object}

\item{...}{arguments to forward to internal functions}
}
\description{
inference of the alternate allele based on count data
}
\details{
The inference essentially ranks all alleles and the most expressed allele not
declared as reference will be inferred as the alternative allele. At the moment only inference of
bi-allelic alternative alleles are available.
}
\examples{
#load data
data(ASEset)

inferAltAllele(ASEset)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{phase}

