% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inferGenotypes}
\alias{inferGenotypes}
\title{infererence of genotypes from ASEset count data}
\usage{
inferGenotypes(x, min.allele.fraction = 0.05, min.infer.count = 10,
  return.class = "list", biallelic.only = TRUE)
}
\arguments{
\item{x}{ASEset}

\item{min.allele.fraction}{min allele fraction}

\item{min.infer.count}{min threshold of counts to infer}

\item{return.class}{'list' or 'array'}

\item{biallelic.only}{logical}
}
\description{
inference of genotypes
}
\details{
Oftern necessary information to link AI to SNPs outside coding region
}
\examples{
data(ASEset)
g <- inferGenotypes(ASEset)
}
\author{
Jesper R. Gadin
}
\keyword{infer}

