% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inferSnp}
\alias{inferSnp}
\title{inference of SNPs of ASEset}
\usage{
inferSnp(x, strand = "*", biallelic.only = TRUE)
}
\arguments{
\item{x}{ASEset}

\item{strand}{strand}

\item{biallelic.only}{logical}
}
\description{
inference of SNPs
}
\details{
Oftern necessary information to link AI to SNPs outside coding region
}
\examples{
data(ASEset)
g <- inferSnp(ASEset)
}
\author{
Jesper R. Gadin
}
\keyword{infer}

