% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initialize-GlobalAnalysis}
\alias{gba}
\alias{initialize-GlobalAnalysis}
\title{Initialize GlobalAnalysis}
\usage{
gba(x = "ASEset", ...)
}
\arguments{
\item{x}{\code{ASEset}}

\item{...}{internal arguments}
}
\description{
Functions to construct GlobalAnalysis objects
}
\details{
produces a class container for a global analysis
}
\examples{
data(ASEset)
a <- ASEset
gba <- gba(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{global}

