% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initialize-ReferenceBias}
\alias{initialize,ReferenceBias-method}
\alias{initialize-ReferenceBias}
\alias{refBias}
\title{Initialize ReferenceBias}
\usage{
\S4method{initialize}{ReferenceBias}(.Object, x = "ASEset")

refBias(x)
}
\arguments{
\item{.Object}{to be ReferenceBias-class object}

\item{x}{\code{ASEset}}
}
\description{
Functions to construct ReferenceBias objects
}
\details{
produces a class container for reference bias calculations
}
\examples{
data(ASEset)
a <- ASEset
genotype(a) <- inferGenotypes(a)
a <- refAllele(a,
 	fasta=system.file('extdata/hg19.chr17.fa',
 	package='AllelicImbalance'))
refbiasObject <- refBias(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{bias}
\keyword{mapbias}
\keyword{refBias}

