% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize-methods.R
\name{initialize-riskVariant}
\alias{initialize-riskVariant}
\alias{riskVariantFromGRanges}
\title{Initialize riskVariant}
\usage{
riskVariantFromGRanges(x, genotype, colData = NULL, ...)
}
\arguments{
\item{x}{GRanges object for the SNPs}

\item{genotype}{matrix}

\item{colData}{A \code{DataFrame} object containing sample specific data}

\item{...}{internal arguments}
}
\description{
Functions to construct riskVariant objects
}
\details{
produces a class container for reference bias calculations
}
\examples{
data(ASEset)
ge <- inferGenotypes(ASEset)
rv <- riskVariantFromGRanges(x=GRvariants, genotype=ge)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{bias}
\keyword{mapbias}
\keyword{refBias}

