% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxillary-functions-visuals.R
\name{legendBarplot}
\alias{legendBarplot}
\title{add legend to AllelicImbalance barplot}
\usage{
legendBarplot(lowerLeftCorner, size, rownames, colnames, boxsize = 1,
  boxspace = 1, fgCol, bgCol, ylegendPos = 1, xlegendPos = 0.96,
  cex = 1)
}
\arguments{
\item{lowerLeftCorner}{position of the plot to add legend to (default c(0,0))}

\item{size}{scale the plot, default is 1}

\item{rownames}{rownames in legend}

\item{colnames}{colnames in legend}

\item{boxsize}{size of each box fill}

\item{boxspace}{space inbetween the box fill}

\item{fgCol}{color for allele1}

\item{bgCol}{color for allele2}

\item{ylegendPos}{placement of the legend within the plot for y}

\item{xlegendPos}{placement of the legend within the plot for x}

\item{cex}{size of legend text}
}
\description{
adds a very customizable legend function for AllelicImbalance barplots.
}
\details{
the function is preferably called from within the AllelicImbalance barplot method.
}
\examples{

#code placeholders
#< create a barplot with legend >
#< add legend >
 
}
\author{
Jesper R. Gadin
}
\keyword{barplot}
\keyword{legend}

