% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkVariantAlmlof-class.R
\docType{class}
\name{LinkVariantAlmlof-class}
\alias{LinkVariantAlmlof-class}
\alias{LinkVariantAlmlof}
\alias{LinkVariantAlmlof-method}
\alias{pvalue}
\alias{pvalue,LinkVariantAlmlof-method}
\title{LinkVariantAlmlof class}
\usage{
pvalue(x, ...)

\S4method{pvalue}{LinkVariantAlmlof}(x)
}
\arguments{
\item{x}{LinkVariantAlmlof object}

\item{...}{pass arguments to internal functions}
}
\description{
Object that holds results from AI detection.
}
\details{
The LinkVariantAlmlof-class contains
}
\examples{

#some code

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{LinkVariantAlmlof}
\keyword{class}
