% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicimbalance.R
\name{log_posterior_allelic_fold_change}
\alias{log_posterior_allelic_fold_change}
\title{log posterior allelic fold change}
\usage{
log_posterior_allelic_fold_change(count1, count2, shape1, shape2,
  level = 0.99)
}
\arguments{
\item{count1, count2}{allele counts}

\item{shape1, shape2}{shape parameters}

\item{level}{confidence level for credible interval}
}
\value{
\describe{
  \item{lpafc}{the effect size.}
  \item{lower}{lower bound for the credible interval.}
  \item{upper}{upper bound for the credible interval.}
}
}
\description{
Bayesian effect size for imbalance data
}
