% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxillary-methods.R
\docType{methods}
\name{lva}
\alias{lva}
\alias{lva,ASEset-method}
\alias{lva,array-method}
\title{lva}
\usage{
lva(x, ...)

\S4method{lva}{ASEset}(x, rv, region, settings = list(),
  return.class = "matrix", return.meta = FALSE, ...)
}
\arguments{
\item{x}{ASEset object with phase and 'ref'/'alt' allele information}

\item{...}{arguments to forward to internal functions}

\item{rv}{riskVariant object with phase and 'ref'/'alt' allele information}

\item{region}{riskVariant object with phase and alternative allele information}

\item{settings}{riskVariant object with phase and alternative allele information}

\item{return.class}{'vector' or 'matrix'}

\item{return.meta}{logical to return a list with metainformation}
}
\description{
make an almlof regression for arrays
}
\details{
internal method that takes one array with results from regionSummary
and one matrix with group information for each risk SNP (based on phase)
}
\examples{
data(ASEset)
a <- ASEset
# Add phase
set.seed(1)
p1 <- matrix(sample(c(1,0),replace=TRUE, size=nrow(a)*ncol(a)),nrow=nrow(a), ncol(a))
p2 <- matrix(sample(c(1,0),replace=TRUE, size=nrow(a)*ncol(a)),nrow=nrow(a), ncol(a))
p <- matrix(paste(p1,sample(c("|","|","/"), size=nrow(a)*ncol(a), replace=TRUE), p2, sep=""),
	nrow=nrow(a), ncol(a))

phase(a) <- p

#add alternative allele information
mcols(a)[["alt"]] <- inferAltAllele(a)

#init risk variants
ge <- inferGenotypes(ASEset)
rv <- riskVariantFromGRanges(x=GRvariants, genotype=ge)
phase(rv) <- p

# in this example all snps in the ASEset defines the region
r1 <- granges(a)

# in this example two overlapping subsets of snps in the ASEset defines the region
r2 <- split(granges(a)[c(1,2,2,3)],c(1,1,2,2))

# use a multilevel list as input (output will keep the list dimensions)
region <- split(granges(a)[c(1,2,2,3)],c(1,1,2,2))
names(region) <- c("introns", "exons")
r3 <- list(g1=list(tx1=region, tx2=region), g2=list(tx1=region, tx2=region, tx3=region))

# link variant almlof (lva)
lva(a, rv, r1)
lva(a, rv, r2)
lva(a, rv, r3)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{phase}

