% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeMapBiasReads}
\alias{makeMapBiasReads}
\title{simulate reads from ASEset}
\usage{
makeMapBiasReads(x, reads = 100, refPath, read.length = 100, PE = FALSE,
  insert.size = -30, insert.sd = 1)
}
\arguments{
\item{x}{ASEset}

\item{reads}{number of reads to create}

\item{refPath}{reference genome path}

\item{read.length}{length of reads}

\item{PE}{paired-end reads, TRUE or FALSE}

\item{insert.size}{Nr bases separating the two reads, only if PE=TRUE}

\item{insert.sd}{standard deviation for the insert.size}
}
\description{
simulate reads
}
\details{
to compensate for reference genome mapping bias
}
\examples{
data(ASEset)
g <- makeMapBiasReads(ASEset)
}
\author{
Jesper R. Gadin
}
\keyword{infer}

