% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxillary-functions.R
\name{makeMaskedFasta}
\alias{makeMaskedFasta}
\title{makes masked fasta reference}
\usage{
makeMaskedFasta(fastaIn, fastaOut, posToReplace, splitOnSeqlevels = TRUE,
  verbose = TRUE)
}
\arguments{
\item{fastaIn}{character string of the path for the fasta file to be used}

\item{fastaOut}{character string of the path for the masked fasta file (no extension)}

\item{posToReplace}{GRanges object with the genomic ranges to replace}

\item{splitOnSeqlevels}{write on file for each seqlevel to save memory}

\item{verbose}{makes function more talkative}
}
\description{
Replaces all selected positions in a fasta file with the character N
}
\examples{
data(ASEset.sim)
gr <- rowRanges(ASEset.sim)
fastaIn <- system.file('extdata/hg19.chr17.subset.fa', package='AllelicImbalance')
makeMaskedFasta(fastaIn=fastaIn, fastaOut="fastaOut",posToReplace=gr)
}
\author{
Jesper R. Gadin
}
\keyword{fasta}
\keyword{masked}
\keyword{reference}

