% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{mapbias}
\alias{mapbias}
\alias{mapbias,ASEset-method}
\title{mapbias effect}
\arguments{
\item{x}{\code{ASEset} object}

\item{strand}{strand option}
}
\description{
measures the effect of mapbias over heterozygous SNPs
}
\details{
The distribution of fractions for the ref allele and the
alternative are compared
}
\examples{
#load example data
data(ASEset)

#this example data contains to few SNPs to actually
#measure the effect reliably, but as example it serves
#the purpose.

mapbias(ASEset)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{mapbias}

