% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{mapBiasRef}
\alias{mapBiasRef}
\alias{mapBiasRef,ASEset-method}
\title{mapBias for reference allele}
\usage{
mapBiasRef(x, ...)

\S4method{mapBiasRef}{ASEset}(x)
}
\arguments{
\item{x}{\code{ASEset} object}

\item{...}{internal arguments}
}
\description{
Create a matrix of bias for the reference allele
}
\details{
select the expected frequency for the reference allele
}
\examples{
#load example data
data(ASEset)
a <- ASEset

mat <- mapBiasRef(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{mapbias}

