% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{minCountFilt}
\alias{minCountFilt}
\alias{minCountFilt,ASEset-method}
\title{minCountFilt methods}
\usage{
minCountFilt(x, ...)

\S4method{minCountFilt}{ASEset}(x, strand = "*", threshold.counts = 1,
  sum = "all", replace.with = "zero", return.class = "ASEset")
}
\arguments{
\item{x}{\code{ASEset} object}

\item{...}{internal param}

\item{strand}{strand to infer from}

\item{threshold.counts}{cutoff for read counts (see details)}

\item{sum}{'each' or 'all'}

\item{replace.with}{only option 'zero'}

\item{return.class}{'ASEset', 'array' or 'matrix'}
}
\description{
filter on minCountFilt snps
}
\details{
Description info here
}
\examples{

#load example data
data(ASEset)
a <- ASEset

minCountFilt(a)

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{filter}
