% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{minFreqFilt}
\alias{minFreqFilt}
\alias{minfreqFilt}
\alias{minFreqFilt,ASEset-method}
\title{minFreqFilt methods}
\usage{
minFreqFilt(x, ...)

\S4method{minFreqFilt}{ASEset}(x, strand = "*",
  threshold.frequency = 0.1, replace.with = "zero",
  return.class = "ASEset", sum = "all")
}
\arguments{
\item{x}{\code{ASEset} object}

\item{...}{internal param}

\item{strand}{strand to infer from}

\item{threshold.frequency}{least fraction to classify (see details)}

\item{replace.with}{only option 'zero'}

\item{return.class}{'ASEset', 'array' or 'matrix'}

\item{sum}{'each' or 'all'}
}
\description{
filter on minFreqFilt snps
}
\details{
Description info here
}
\examples{

#load example data
data(ASEset)
a <- ASEset

minFreqFilt(a)

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{filter}
