% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{phaseArray2Matrix}
\alias{phaseArray2Matrix}
\alias{phaseArray2Matrix,array-method}
\title{phaseArray2Matrix}
\usage{
phaseArray2Matrix(x, ...)

\S4method{phaseArray2Matrix}{array}(x, ...)
}
\arguments{
\item{x}{array see examples}

\item{...}{arguments to forward to internal functions}
}
\description{
used to convert the phase from the visually friendly matrix to array.
}
\details{
A more effectice way of store the phase data in the ASEset object
}
\examples{
#load data
data(ASEset)
a <- ASEset

#example phase matrix
p1 <- matrix(sample(c(1,0),replace=TRUE, size=nrow(a)*ncol(a)),nrow=nrow(a), ncol(a))
p2 <- matrix(sample(c(1,0),replace=TRUE, size=nrow(a)*ncol(a)),nrow=nrow(a), ncol(a))
p <- matrix(paste(p1,sample(c("|","|","/"), size=nrow(a)*ncol(a), replace=TRUE), p2, sep=""),
	nrow=nrow(a), ncol(a))

ar <- phaseMatrix2Array(p)

#Convert back
mat <- phaseArray2Matrix(ar)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{phase}

