% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapbias-methods.R
\docType{methods}
\name{randomRef}
\alias{randomRef}
\alias{randomRef,ASEset-method}
\title{Random ref allele from genotype}
\usage{
randomRef(x, ...)

\S4method{randomRef}{ASEset}(x, inferGenotypes = FALSE)
}
\arguments{
\item{x}{\code{ASEset} object}

\item{...}{internal arguments}

\item{inferGenotypes}{infer genotypes from count matrix}
}
\description{
Create a vector of random reference alleles
}
\details{
Randomly shuffles which of the two alleles for each genotype that is
indicated as reference allele.
}
\examples{
#load example data
data(ASEset.sim)
a <- ASEset.sim

mcols(a)[["ref"]] <- randomRef(a, inferGenotypes=TRUE)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{mapbias}

