\name{reads}
\docType{data}
\alias{reads}

\title{reads object}
\description{
	This data set corresponds to the BAM-file data import illustrated in the vignette. The data set consists of a chromosome 17 region from 20 RNA-seq experiments of HapMap samples.
}


\author{Jesper R. Gadin, Lasse Folkersen}
\seealso{
	\itemize{
		\item  The \code{\link{GRvariants}} which is another example object
	}
}
\references{
	Montgomery SB et al. Transcriptome genetics using second generation sequencing in a Caucasian population. Nature. 2010 Apr 1;464(7289):773-7.
}
\examples{
	##load eample data (Not Run)  
	#data(reads)
}
\keyword{data}
\keyword{object}
\keyword{example}

