% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{refAllele}
\alias{refAllele}
\alias{refAllele,ASEset-method}
\title{Reference allele}
\usage{
refAllele(x, fasta)
}
\arguments{
\item{x}{\code{ASEset} object}

\item{strand}{strand option}
}
\description{
Extract the allele based on SNP location from the reference fasta file
}
\details{
The alleles will be placed in the rowData() meta column 'ref'
}
\examples{
#load example data
data(ASEset)

fasta <- system.file('extdata/hg19.subset.fa', package='AllelicImbalance')
refAllele(ASEset)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{mapbias}
\keyword{reference}

