% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ReferenceBias-class}
\alias{ReferenceBias}
\alias{ReferenceBias-class}
\title{ReferenceBias class}
\arguments{
\item{x}{ReferenceBias object}

\item{strand}{which strand of '+', '-' or '*'}

\item{verbose}{makes function more talkative}
}
\value{
An object of class ReferenceBias storing reference fractions.
}
\description{
Object that holds Reference bias information
}
\details{
Reference bias-class contain annotated fractions of the reference allele

Shortly the RefBias-class most prominent purpose is to be able to dispatch on
methods like, 'plot', 'table', 'summary' and similar.
}
\section{Constructor}{
 refBias(x = ASEset)

\describe{

Arguments: \item{x}{an ASEset object  }

}
}
\examples{
data(ASEset)
refbiasObject <- refBiad(ASEset)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{ASEset}
\keyword{class}

