% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{ReferenceBias-show}
\alias{ReferenceBias-show}
\alias{ReferenceBias-show,ReferenceBias-method}
\alias{show,ReferenceBias-method}
\title{genotype filter methods}
\usage{
\S4method{show}{ReferenceBias}(object)
}
\arguments{
\item{object}{ReferenceBias object}
}
\description{
useful genotype filters
}
\details{
These filters are called upon ASEset objects
}
\examples{
#load example data
cat("there is no example data yet for this method")
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{show}

