% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{refFraction}
\alias{refFraction}
\alias{refFraction,ASEset-method}
\title{reference fraction}
\usage{
refFraction(x, strand = "*")
}
\arguments{
\item{x}{\code{ASEset} object}

\item{strand}{strand option}
}
\description{
The fractions for all heterozygote reference alleles
}
\details{
Neccessary to measure the effect of mapbias over heterozygous SNPs
}
\examples{
#load example data
data(ASEset)
a <- ASEset

#this example data contains to few SNPs to actually
#measure the effect reliably, but as example it serves
#the purpose.

#prepare ASEset
genotype(a) <- inferGenotypes(a)
a <- refAllele(a,
		fasta=system.file('extdata/hg19.chr17.fa',
		package='AllelicImbalance'))

rf <- refFraction(a, strand="*")
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{fraction}
\keyword{reference}

