% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegionSummary-class.R
\docType{class}
\name{RegionSummary-class}
\alias{RegionSummary}
\alias{RegionSummary-class}
\alias{RegionSummary-method}
\title{RegionSummary class}
\arguments{
\item{value}{argument used for replacement}

\item{...}{pass arguments to internal functions}
}
\description{
Object that holds results from AI detection.
}
\details{
The RegionSummary-class contains
}
\examples{
#some code
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{RegionSummary}
\keyword{class}

