% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxillary-methods.R
\docType{methods}
\name{regionSummary}
\alias{regionSummary}
\alias{regionSummary,ASEset-method}
\alias{regionSummary,numeric-method}
\title{regionSummary}
\usage{
regionSummary(x, ...)

\S4method{regionSummary}{ASEset}(x, gr, strand = "*", ...)
}
\arguments{
\item{x}{ASEset object}

\item{...}{arguments to forward to internal functions}

\item{gr}{GenomicRanges object to summmarize over}

\item{strand}{can be "+", "-" or "*"}
}
\description{
Gives a summary of AI-consistency for a transcript
}
\details{
From a given set of e.g. transcripts exon ranges the function will return
a summary for the sum of all exons. Phase information is required.
}
\examples{
data(ASEset)
a <- ASEset
# Add phase
set.seed(1)
p1 <- matrix(sample(c(1,0),replace=TRUE, size=nrow(a)*ncol(a)),nrow=nrow(a), ncol(a))
p2 <- matrix(sample(c(1,0),replace=TRUE, size=nrow(a)*ncol(a)),nrow=nrow(a), ncol(a))
p <- matrix(paste(p1,sample(c("|","|","/"), size=nrow(a)*ncol(a), replace=TRUE), p2, sep=""),
	nrow=nrow(a), ncol(a))

phase(a) <- p

#add alternative allele information
mcols(a)[["alt"]] <- inferAltAllele(a)

# in this example every snp is on its own exon
txGR <- granges(a)
t <- regionSummary(a, txGR)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{summary}

