% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/riskVariant-class.R
\docType{class}
\name{riskVariant-class}
\alias{alt,riskVariant-method}
\alias{alt<-,riskVariant,ANY-method}
\alias{phase,riskVariant-method}
\alias{phase<-,riskVariant-method}
\alias{ref,riskVariant-method}
\alias{ref<-,riskVariant,ANY-method}
\alias{riskVariant}
\alias{riskVariant-class}
\alias{riskVariant-method}
\title{riskVariant class}
\usage{
\S4method{ref}{riskVariant}(x)

\S4method{ref}{riskVariant,ANY}(x) <- value

\S4method{alt}{riskVariant}(x)

\S4method{alt}{riskVariant,ANY}(x) <- value

\S4method{phase}{riskVariant}(x, return.class = "matrix")

\S4method{phase}{riskVariant}(x) <- value
}
\arguments{
\item{x}{riskVariant object or list of riskVariants}

\item{value}{argument used for replacement}

\item{return.class}{type of class returned eg. "list or ""array".}

\item{...}{pass arguments to internal functions}
}
\description{
Object that holds results from AI detection.
}
\details{
The riskVariant-class contains
}
\examples{
#some code
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{class}
\keyword{riskVariant}

