% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxillary-methods-annotation.R
\docType{methods}
\name{txToArray}
\alias{txToArray}
\alias{txToArray,ASEset-method}
\alias{txToArray,character-method}
\title{transform to array}
\usage{
txToArray(pathBam, ...)

\S4method{txToArray}{character}(pathBam, pathVcf, pathGFF = NULL, verbose)
}
\arguments{
\item{...}{arguments to pass on}

\item{pathGFF}{path to gff file}

\item{verbose}{makes function more talkative}
}
\description{
transform ranges to compatible ASEset array
}
\examples{
#generate example ASEset
snps <- 1000
samples <-5
ar <-array(rep(unlist(lapply(1:snps,
			function(x){(sample(c(TRUE,FALSE,TRUE,FALSE), size = 4))})), samples),
			dim=c(4,snps,samples))
ar2 <- array(sample(50:300, 4*snps*samples,replace=TRUE), dim=c(4,snps,samples))
ar2[ar] <- 0
ar2 <- aperm(ar2, c(2, 3, 1))

#make example rowRanges for the corresponding information
gr <- GRanges(
	seqnames = Rle(sample(paste("chr", 1:3,sep=""),snps, replace=TRUE)),
	ranges = IRanges(1:snps, width = 1, names= paste("snp",1:snps,sep="")),
	strand="*"
)
#make ASEset from array
a <- ASEsetFromArrays(gr, countsUnknown=ar2)

# gene annotation
geneGR <- GRanges(
 	seqnames = Rle(sample(paste("chr", 1:3,sep=""),snps, replace=TRUE)),
 	ranges = IRanges(1:snps, width = 1, names= paste("snp",1:snps,sep="")),
 	strand="*"
)

# tx annotation
geneGR <- GRanges(
 	seqnames = Rle(sample(paste("chr",1:22,sep=""),snps, replace=TRUE)),
 	ranges = IRanges(1:snps, width = 1, names= paste("snp",1:snps,sep="")),
 	strand="*"
)
}
\author{
Jesper R. Gadin
}
\keyword{array}

