% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-methods.R
\docType{methods}
\name{ASEset-filters}
\alias{ASEset-filters}
\alias{hetFilt}
\alias{hetFilt,ASEset-method}
\alias{minCountFilt}
\alias{minCountFilt,ASEset-method}
\alias{minFreqFilt}
\alias{minFreqFilt,ASEset-method}
\alias{multiAllelicFilt}
\alias{multiAllelicFilt,ASEset-method}
\title{genotype filter methods}
\usage{
hetFilt(x, ...)

\S4method{hetFilt}{ASEset}(x, source = "genotype", ...)

multiAllelicFilt(x, ...)

\S4method{multiAllelicFilt}{ASEset}(x, strand = "*",
  threshold.count.sample = 10, threshold.frequency = 0.1,
  filterOver = "eachSample")

minFreqFilt(x, ...)

\S4method{minFreqFilt}{ASEset}(x, strand = "*",
  VariantAlleleSource = "alleleCounts", threshold.frequency = 0.1)

minCountFilt(x, ...)

\S4method{minCountFilt}{ASEset}(x, strand = "*", threshold.counts = 1,
  sum = "all", replace.with = "zero", return.class = "ASEset")
}
\arguments{
\item{x}{ASEset object}

\item{source}{'genotype' or 'alleleCounts'}
}
\description{
useful genotype filters
}
\details{
hetFilt returns TRUE if the samples is heterozygote, based on stored genotype information
present in the phase data.
}
\examples{

#load example data
data(ASEset)
a <- ASEset

genotype(a) <- inferGenotypes(a)
hets <- hetFilt(a) 

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{filter}

