% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{binom.test}
\alias{binom.test}
\alias{binom.test,ASEset-method}
\title{binomial test}
\usage{
binom.test(x, n, p = 0.5, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95)

\S4method{binom.test}{ASEset}(x, n = "nonStranded")
}
\arguments{
\item{x}{\code{ASEset} object}

\item{n}{strand option}

\item{p}{hypothesized probability of success.}

\item{alternative}{indicates the alternative hypothesis and must be
    one of \code{"two.sided"}, \code{"greater"} or \code{"less"}.
    You can specify just the initial letter.}

\item{conf.level}{confidence level for the returned confidence
    interval.}
}
\value{
\code{binom.test} returns a matrix
}
\description{
Performs a binomial test on an ASEset object.
}
\details{
the test can only be applied to one strand at the time.
}
\examples{
\\t#load example data
\\tdata(ASEset)
\\t
\\t#make a binomial test
\\tbinom.test(ASEset,'nonStranded')
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The \code{\link{chisq.test}} which is another test
that can be applied on an \link{ASEset} object.  }
}
\keyword{binomial}
\keyword{test}

