% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{boxplot}
\alias{boxplot}
\alias{boxplot,ReferenceBias-method}
\title{boxplots}
\usage{
\S4method{boxplot}{ReferenceBias}(x, strand = "*", labels.axis = TRUE, ...)
}
\arguments{
\item{x}{\code{ReferenceBias} object}

\item{strand}{'+','-' or '*'}

\item{labels.axis}{logical}

\item{...}{arguments to forward to internal boxplots function}
}
\description{
uses base graphics box plot
}
\details{
The boxplot will show the density over frequencies for each sample
}
\examples{
#load example data

data("ReferenceBias")
boxplot(ReferenceBias)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{box}
\keyword{plot}

