% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{chisq.test}
\alias{chisq.test}
\alias{chisq.test,ASEset-method}
\title{chi-square test}
\usage{
chisq.test(x, y = NULL, correct = TRUE, p = rep(1/length(x), length(x)),
  rescale.p = FALSE, simulate.p.value = FALSE, B = 2000)

\S4method{chisq.test}{ASEset}(x, y = "nonStranded")
}
\arguments{
\item{x}{\code{ASEset} object}

\item{y}{strand option}

\item{correct}{a logical indicating whether to apply continuity
    correction when computing the test statistic for 2 by 2 tables: one
    half is subtracted from all \eqn{|O - E|} differences; however, the
    correction will not be bigger than the differences themselves.  No correction
    is done if \code{simulate.p.value = TRUE}.}

\item{p}{a vector of probabilities of the same length of \code{x}.
    An error is given if any entry of \code{p} is negative.}

\item{rescale.p}{a logical scalar; if TRUE then \code{p} is rescaled
    (if necessary) to sum to 1.  If \code{rescale.p} is FALSE, and
    \code{p} does not sum to 1, an error is given.}

\item{simulate.p.value}{a logical indicating whether to compute
    p-values by Monte Carlo simulation.}

\item{B}{an integer specifying the number of replicates used in the
    Monte Carlo test.}
}
\value{
\code{chisq.test} returns a matrix with the chisq.test P-value for
each SNP and sample
}
\description{
Performs a chisq.test on an ASEset object.
}
\details{
The test is performed on one strand in an ASEset object.
}
\examples{
#load example data
data(ASEset)

#make a chi-square test on default nonStranded strand
chisq.test(ASEset)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The \code{\link{binom.test}} which is another test
that can be applied on an \link{ASEset} object.  }
}
\keyword{chi-square}
\keyword{test}

