% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{chisq.test}
\alias{chisq.test}
\alias{chisq.test,ASEset-method}
\title{chi-square test}
\usage{
chisq.test(x, y = NULL, correct = TRUE, p = rep(1/length(x), length(x)),
  rescale.p = FALSE, simulate.p.value = FALSE, B = 2000)

\S4method{chisq.test}{ASEset}(x, y = "*")
}
\arguments{
\item{x}{\code{ASEset} object}

\item{y}{strand option}

\item{correct}{NOT USED}

\item{p}{NOT USED}

\item{rescale.p}{NOT USED}

\item{simulate.p.value}{NOT USED}

\item{B}{NOT USED}
}
\value{
\code{chisq.test} returns a matrix with the chisq.test P-value for
each SNP and sample
}
\description{
Performs a chisq.test on an ASEset object.
}
\details{
The test is performed on one strand in an ASEset object.
}
\examples{
#load example data
data(ASEset)

#make a chi-square test on default non-stranded strand
chisq.test(ASEset)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The \code{\link{binom.test}} which is another test
that can be applied on an \link{ASEset} object.  }
}
\keyword{chi-square}
\keyword{test}

