% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coverageMatrixListFromGAL}
\alias{coverageMatrixListFromGAL}
\title{coverage matrix of GAlignmentsList}
\usage{
coverageMatrixListFromGAL(BamList, strand = NULL,
  ignore.empty.bam.row = TRUE)
}
\arguments{
\item{BamList}{GAlignmentsList containing reads over the region to calculate
coverage}

\item{strand}{strand has to be '+' or '-'}

\item{ignore.empty.bam.row}{argument not in use atm}
}
\description{
Get coverage per nucleotide for reads covering a region
}
\details{
a convenience function to get the coverage from a list of reads stored in
GAlignmnetsList, and returns by default a list with one matrix, and
information about the genomic start and stop positions.
}
\examples{
\\tr <- reads
\\tseqlevels(r) <- '17'
\\tcovMatList <- coverageMatrixListFromGAL(BamList=r, strand='+')
}
\author{
Jesper R. Gadin
}
\keyword{coverage}

