% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectedAI-class.R
\docType{class}
\name{DetectedAI-class}
\alias{DetectedAI-class}
\alias{DetectedAI}
\alias{DetectedAI-method}
\alias{referenceFrequency}
\alias{referenceFrequency,DetectedAI-method}
\alias{thresholdFrequency}
\alias{thresholdFrequency,DetectedAI-method}
\alias{thresholdCountSample}
\alias{thresholdCountSample,DetectedAI-method}
\alias{thresholdDeltaFrequency}
\alias{thresholdDeltaFrequency,DetectedAI-method}
\alias{thresholdPvalue}
\alias{thresholdPvalue,DetectedAI-method}
\title{DetectedAI class}
\usage{
referenceFrequency(x, ...)

\S4method{referenceFrequency}{DetectedAI}(x, return.class = "array")

thresholdFrequency(x, ...)

\S4method{thresholdFrequency}{DetectedAI}(x, return.class = "array")

thresholdCountSample(x, ...)

\S4method{thresholdCountSample}{DetectedAI}(x, return.class = "array")

thresholdDeltaFrequency(x, ...)

\S4method{thresholdDeltaFrequency}{DetectedAI}(x, return.class = "array")

thresholdPvalue(x, ...)

\S4method{thresholdPvalue}{DetectedAI}(x, return.class = "array")
}
\arguments{
\item{x}{ASEset object or list of ASEsets}

\item{...}{pass arguments to internal functions}

\item{return.class}{type of class returned eg. "list or ""array".}
}
\description{
Object that holds results from AI detection.
}
\details{
The DetectedAI-class contains
}
\examples{

data(ASEset)
a <- ASEset
dai <- detectAI(a)

#summary(gba)
#write.tables(dai)

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{ASEset}
\keyword{class}
