% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxillary-functions.R
\name{gba}
\alias{gba}
\title{global analysis wrapper}
\usage{
gba(pathBam, pathVcf, pathGFF = NULL, verbose)
}
\arguments{
\item{pathBam}{path to bam file}

\item{pathVcf}{path to vcf file}

\item{pathGFF}{path to gff file}

\item{verbose}{makes function more talkative}
}
\description{
A wrapper to make a global analysis based on paths for BAM, VCF and GFF files
}
\examples{
#empty as function doesn't exist
}
\author{
Jesper R. Gadin
}
\keyword{global}
\keyword{wrapper}

