% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{GRvariants}
\alias{GRvariants}
\title{GRvariants object}
\description{
this data is a \code{GRanges} object that contains the ranges for three
example SNPs.
}
\examples{
\\t#load example data
\\tdata(GRvariants)
\\t
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The \code{\link{reads}} which is another example
object }
}
\keyword{data}
\keyword{example}
\keyword{object}

