\name{GRvariants}
\docType{data}
\alias{GRvariants}

\title{GRvariants object}
\description{
	this data is a \code{GRanges} object that contains the ranges for three example SNPs.
}


\author{Jesper R. Gadin, Lasse Folkersen}
\seealso{
	\itemize{
		\item  The \code{\link{reads}} which is another example object
	}
}
\examples{
	#load example data
	data(GRvariants)
	
}
\keyword{data}
\keyword{object}
\keyword{example}


