% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histplot-methods.R
\docType{methods}
\name{histplot}
\alias{histplot}
\alias{hist}
\alias{hist,ReferenceBias-method}
\alias{hist,ASEset-method}
\title{histogram plots}
\usage{
\S4method{hist}{ASEset}(x, strand = "*", type = "mean", log = 1, ...)
}
\arguments{
\item{x}{\code{ReferenceBias} object or \code{ASEset} object}

\item{strand}{'+','-' or '*'}

\item{type}{'mean' (only one option atm)}

\item{log}{an integer to log each value (integer 10 for log10)}

\item{...}{arguments to forward to interal boxplots function}
}
\description{
uses base graphics hist plot
}
\details{
The histogram will show the density over frequencies for each sample
}
\examples{

##load example data

#data(ASEset)
#a <- ASEset
#hist(a)

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{hist}
\keyword{plot}
