% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auxillary-functions.R
\name{import-bam}
\alias{impBamGAL}
\alias{impBamGRL}
\alias{import-bam}
\title{Import Bam}
\usage{
impBamGRL(UserDir, searchArea, readLength = 150, reduce = FALSE,
  scanBamFlag = NULL, verbose = TRUE)

impBamGAL(UserDir, searchArea, readLength = 150, reduce = FALSE,
  scanBamFlag = NULL, XStag = FALSE, verbose = TRUE)
}
\arguments{
\item{UserDir}{The relative or full path of folder containing bam files.}

\item{searchArea}{A \code{GenomicRanges object} that contains the regions of
interest}

\item{readLength}{length of sequence reads (default: 150)}

\item{reduce}{booleon to hinder importing same read multiple times
(see the details section)}

\item{scanBamFlag}{set scanBamFlag using scanBamFlag()}

\item{verbose}{Setting \code{verbose=TRUE} gives details of procedure during
function run.}

\item{XStag}{Setting \code{XStag=TRUE}, if the XS-tag is present the strand
specific information will be placed in the mcols slot 'XS'}
}
\value{
\code{impBamGRL} returns a GRangesList object containing the RNA-seq
reads in the region defined by the \code{searchArea} argument.
\code{impBamGAL} returns a list with GAlignments objects containing the
RNA-seq reads in the region defined by the \code{searchArea} argument.
\code{funImpBamGAPL} returns a list with GappedAlignmentPairs object
containing the RNA-seq reads in the region defined by the \code{searchArea}
argument.
}
\description{
Imports a specified genomic region from a bam file using a GenomicRanges
object as search area.
}
\details{
These functions are wrappers to import bam files into R and store them into
either GRanges, GAlignments or GappedAlignmentpairs objects.

It is recommended to use the impBamGAL() which takes information of gaps
into account. If searchArea consists of overlapping ranges or ranges with
an inter-distance shorter than the read length some reads may be imported
multiple times. To avoid this set option reduce=TRUE and if the searchArea
symbolises SNPs, then set readLength to the length of reads in sequencing.

If the sequence data is strand-specific you may want to set XStag=TRUE. The
strand specific information will then be stored in the meta columns with
column name 'XS'.
}
\note{
A typical next step after the import of bam files is to obtain SNP
information. This can be done either with the \code{\link{impBcfGRL}} and
\code{\link{getAlleleCounts}} functions. Alternatively the
\code{\link{scanForHeterozygotes}} function provides R-based functionality
for identifying heterozygote coding SNPs.
}
\examples{
#Declare searchArea
searchArea <- GRanges(seqnames=c('17'), ranges=IRanges(79478301,79478361))

#Relative or full path
pathToFiles <- system.file('extdata/ERP000101_subset',
									package='AllelicImbalance')

reads <- impBamGAL(pathToFiles,searchArea,verbose=FALSE)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\seealso{
\itemize{ \item The \code{\link{impBcfGRL}} for importing Bcf
files.  }
}
\keyword{bam}
\keyword{import}

