% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{inferAlleles}
\alias{inferAlleles}
\alias{inferAlleles,ASEset-method}
\title{inference of SNPs of ASEset}
\usage{
inferAlleles(x, strand = "*", return.type = "all",
  threshold.frequency = 0.2, threshold.count.sample = 5,
  inferOver = "eachSample", allow.NA = FALSE)
}
\arguments{
\item{x}{ASEset}

\item{strand}{strand to infer from}

\item{inferOver}{'eachSample' or 'allSamples'}

\item{allow.NA}{treat NA as zero when TRUE}

\item{return.type}{'uni' 'bi' 'tri' 'quad' 'all'}

\item{threshold.count.sample}{least amount of counts to try to infer allele}

\item{threshold.frequency}{least fraction to classify (see details)}
}
\description{
inference of SNPs
}
\details{
threshold.frequency is the least fraction needed to classify as bi tri or
quad allelic SNPs. if 'all' then all of bi tri and quad allelic SNPs will use the same
threshold. Everything under the treshold will be regarded as noise. 'all' will return

a matrix with snps as rows and uni bi tri and quad will be columns. #'
}
\examples{
data(ASEset)
i <- inferAlleles(ASEset)
}
\author{
Jesper R. Gadin
}
\keyword{infer}

