% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inferGenotypes}
\alias{inferGenotypes}
\alias{inferGenotypes,ASEset-method}
\title{infererence of genotypes from ASEset count data}
\usage{
inferGenotypes(x, strand = "*", return.class = "matrix",
  return.allele.allowed = c("bi", "tri", "quad"), threshold.frequency = 0.2,
  threshold.count.sample = 0)
}
\arguments{
\item{x}{ASEset}

\item{strand}{strand to infer from}

\item{threshold.count.sample}{least amount of counts to try to infer allele}

\item{threshold.frequency}{least fraction to classify (see details)}

\item{return.allele.allowed}{vector with 'bi' 'tri' or 'quad'.
'uni' Always gets returned}

\item{return.class}{'matrix' or 'vector'}
}
\description{
inference of genotypes
}
\details{
Oftern necessary information to link AI to SNPs outside coding region
}
\examples{
data(ASEset)
g <- inferGenotypes(ASEset)
}
\author{
Jesper R. Gadin
}
\keyword{infer}

