% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/initialize-methods.R
\name{initialize-GlobalAnalysis}
\alias{GAnalysis}
\alias{initialize-GlobalAnalysis}
\title{Initialize GlobalAnalysis}
\usage{
GAnalysis(x = "ASEset", ...)
}
\arguments{
\item{x}{\code{ASEset}}

\item{...}{internal arguments}
}
\description{
Functions to construct GlobalAnalysis objects
}
\details{
produces a class container for a global analysis
}
\examples{
data(ASEset)
a <- ASEset
# gba <- gba(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{global}

