% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initialize-ReferenceBias}
\alias{RBias}
\alias{initialize-ReferenceBias}
\title{Initialize ReferenceBias}
\usage{
RBias(x = "ASEset", ...)
}
\arguments{
\item{x}{\code{ASEset}}

\item{...}{internal arguments}
}
\description{
Functions to construct ReferenceBias objects
}
\details{
produces a class container for reference bias calculations
}
\examples{
data(ASEset)
a <- ASEset
refbiasObject <- RBias(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{bias}
\keyword{mapbias}
\keyword{refBias}

