% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize-methods.R
\name{initialize-riskVariant}
\alias{initialize-riskVariant}
\alias{riskVariantFromGRangesAndPhaseArray}
\title{Initialize riskVariant}
\usage{
riskVariantFromGRangesAndPhaseArray(x, phase, ...)
}
\arguments{
\item{x}{GRanges object for the SNPs}

\item{phase}{array with phaseinfo}

\item{...}{internal arguments}
}
\description{
Functions to construct riskVariant objects
}
\details{
produces a class container for reference bias calculations
}
\examples{
data(ASEset)
#p <- getPhaseFromSomewhere
#rv <- riskVariantFromGRangesAndPhaseArray(x=GRvariants, phase=p)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{bias}
\keyword{mapbias}
\keyword{refBias}

