% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize-methods.R
\name{initialize-RiskVariant}
\alias{RiskVariantFromGRangesAndPhaseArray}
\alias{initialize-RiskVariant}
\title{Initialize RiskVariant}
\usage{
RiskVariantFromGRangesAndPhaseArray(x, phase, ...)
}
\arguments{
\item{x}{GRanges object for the SNPs}

\item{phase}{array with phaseinfo}

\item{...}{internal arguments}
}
\description{
Functions to construct RiskVariant objects
}
\details{
produces a class container for reference bias calculations
}
\examples{

data(ASEset)
#p <- getPhaseFromSomewhere
#rv <- RiskVariantFromGRangesAndPhaseArray(x=GRvariants, phase=p)


}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{bias}
\keyword{mapbias}
\keyword{refBias}

