% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ReferenceBias-class}
\alias{ReferenceBias}
\alias{ReferenceBias-class}
\alias{ReferenceBias-method}
\alias{frequency,ReferenceBias-method}
\alias{hetPerSample}
\alias{hetPerSample,ReferenceBias-method}
\alias{hetPerSnp}
\alias{hetPerSnp,ReferenceBias-method}
\alias{refbiasByAnnotation}
\alias{refbiasByAnnotation,ReferenceBias-method}
\title{ReferenceBias class}
\usage{
refbiasByAnnotation(x, ...)

\S4method{refbiasByAnnotation}{ReferenceBias}(x, TxDb)

\S4method{frequency}{ReferenceBias}(x, return.class = "matrix",
  strand = "*")

hetPerSample(x, ...)

\S4method{hetPerSample}{ReferenceBias}(x, return.class = "vector",
  strand = "*")

hetPerSnp(x, ...)

\S4method{hetPerSnp}{ReferenceBias}(x, return.class = "vector",
  strand = "*")
}
\arguments{
\item{x}{ReferenceBias object}

\item{...}{pass arguments to internal functions}

\item{TxDb}{A \code{transcriptDb} object}

\item{return.class}{class to return}

\item{strand}{'+','-' or '*'}
}
\value{
An object of class ReferenceBias storing reference fractions.
}
\description{
Object that holds Reference bias information
}
\details{
Reference bias-class contain annotated fractions of the reference allele

Shortly the ReferenceBias-class most prominent purpose is quikly to be able to dispatch on
methods like, 'plot', 'summary' and similar.
}
\section{Constructor}{
 refBias(x = ASEset)

\describe{

Arguments: \item{x}{an ASEset object  }

}
}
\examples{
data(ASEset)
a <- ASEset
rb <- RBias(a)
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{ASEset}
\keyword{class}

