% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapbias-methods.R
\docType{methods}
\name{refFraction}
\alias{refFraction}
\alias{refFraction,ASEset-method}
\title{reference fraction}
\usage{
refFraction(x, ...)
}
\arguments{
\item{x}{\code{ASEset} object}

\item{...}{arguments to forward to internal functions}
}
\description{
The fractions for all heterozygote reference alleles
}
\details{
Neccessary to measure the effect of mapbias over heterozygous SNPs
}
\examples{
#load example data
data(ASEset)
a <- ASEset
rf <- refFraction(a, strand="*")
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{fraction}
\keyword{reference}

