% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RegionSummary-class.R
\docType{class}
\name{RegionSummary-class}
\alias{RegionSummary}
\alias{RegionSummary-class}
\alias{RegionSummary-method}
\title{RegionSummary class}
\arguments{
\item{x}{RegionSummary object}

\item{...}{pass arguments to internal functions}
}
\description{
Object that holds results from the regionSummary method
}
\details{
The RegionSummary-class objects contains summaries for specified regions
}
\examples{
#some code
}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{RegionSummary}
\keyword{class}

