% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSummary-class.R
\docType{class}
\name{RegionSummary-class}
\alias{RegionSummary-class}
\alias{RegionSummary}
\alias{RegionSummary-method}
\alias{sumnames}
\alias{sumnames,RegionSummary-method}
\alias{basic}
\alias{basic,RegionSummary-method}
\title{RegionSummary class}
\usage{
sumnames(x, ...)

\S4method{sumnames}{RegionSummary}(x)

basic(x, ...)

\S4method{basic}{RegionSummary}(x)
}
\arguments{
\item{x}{RegionSummary object}

\item{...}{pass arguments to internal functions}
}
\description{
Object that holds results from the regionSummary method
}
\details{
The RegionSummary-class objects contains summaries for specified regions
}
\examples{

#some code

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{RegionSummary}
\keyword{class}
