% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskVariant-class.R
\docType{class}
\name{RiskVariant-class}
\alias{RiskVariant}
\alias{RiskVariant-class}
\alias{RiskVariant-method}
\alias{alt,RiskVariant-method}
\alias{alt<-,RiskVariant,ANY-method}
\alias{phase,RiskVariant-method}
\alias{phase<-,RiskVariant-method}
\alias{ref,RiskVariant-method}
\alias{ref<-,RiskVariant,ANY-method}
\title{RiskVariant class}
\usage{
\S4method{ref}{RiskVariant}(x)

\S4method{ref}{RiskVariant,ANY}(x) <- value

\S4method{alt}{RiskVariant}(x)

\S4method{alt}{RiskVariant,ANY}(x) <- value

\S4method{phase}{RiskVariant}(x, return.class = "matrix")

\S4method{phase}{RiskVariant}(x) <- value
}
\arguments{
\item{x}{RiskVariant object or list of RiskVariants}

\item{value}{argument used for replacement}

\item{return.class}{type of class returned eg. "list or ""array".}

\item{...}{pass arguments to internal functions}
}
\description{
Object that holds results from AI detection.
}
\details{
The RiskVariant-class contains
}
\examples{

#some code

}
\author{
Jesper R. Gadin, Lasse Folkersen
}
\keyword{RiskVariant}
\keyword{class}

